requirements.txt
python-dotenv
google-generativeai
langchain-google-genai
langchain
langchain_community
ollama
beautifulsoup4
PyPDF2
faiss-cpu
streamlit
-e .